test(rendering): sposta i test dell'estrazione fuori da test_score_visualizer
Dieci classi (~500 righe) verificavano envelope_extractor costruendo un
ScoreVisualizer intero per chiamarne il metodo delegante
_get_stream_envelopes. Ora interrogano get_stream_envelopes direttamente:
niente matplotlib, niente Generator, niente figura da chiudere.

test_envelope_extractor.py passa da 11 a 63 test raccolti,
test_score_visualizer.py da 181 a 129: resta il disegno (legenda,
pannelli, display range, autozoom, magnifier, font scale).

test_plot_envelope_keys_is_the_color_universe non si e' spostato: verifica
che la config del visualizer sia alimentata dalla palette, quindi e' un
test del visualizer. Resta li' come TestEnvelopeColorsConfig.

La fixture make_stream si porta dietro i `del` sugli attributi non
valorizzati: non servono piu' a forzare hasattr=False, ma su un MagicMock
quegli attributi esisterebbero come Mock e num_voices in particolare
verrebbe interrogato per la finestra attiva delle voci.

Nota sulla verifica: contando i `def test_` i due file ne avevano 192
prima e 192 dopo, ma pytest ne raccoglieva 189. Due classi
TestBaseParamName nello stesso modulo — la migrata ombreggiava quella
preesistente e tre test sparivano dalla collection senza che niente
diventasse rosso. Rinominata in TestBaseParamNameWithFilter.

Suite 5145 verde, rete di caratterizzazione verde e baseline invariata.
